Added an "hadjust" XML attribute to title block cells, allowing too long texts to be reduced until they fit the cell.
git-svn-id: svn+ssh://svn.tuxfamily.org/svnroot/qet/qet/branches/0.3@1150 bfdf4180-ca20-0410-9c96-a3a8aa849046

+++ sources/qetapp.h
@@ -93,7 +93,7 @@ class QETApp : public QETSingleApplication {
 	public:
 	static void overrideLangDir(const QString &);
 	static QString lang_dir; ///< Dossier contenant les fichiers de langue
-	static QFont diagramTextsFont(int = -1);
+	static QFont diagramTextsFont(qreal = -1.0);
 	static QETDiagramEditor *diagramEditorForFile(const QString &);
 	static QList<QETDiagramEditor *> diagramEditors();
 	static QList<QETElementEditor *> elementEditors();
